refactor: move from actions to /access
